[MOSIP-31232]Renamed files , tested with local docker image
