--- certalert/check/bootstrap.properties
@@ -1,9 +1,0 @@
-[Database]
-db-host = postgres.soil.mosip.net
-db-port = 5432
-db-su-user = postgres
-postgres-password = CWxa42NDdk
-[API]
-mosip-api-internal-host=api-internal.soil.mosip.net
-mosip_deployment_client_secret=j15svk765HXCSyZM
-pre-expiry-days=7
