--- certalert/check/partner.properties
@@ -1,1 +1,0 @@
-PARTNER_ID=mpartner-default-auth,mpartner-default-resident
