--- certalert/renew/bootstrap.properties
@@ -1,9 +1,0 @@
-[Database]
-db-host = postgres.dev1.mosip.net
-db-port = 5432
-db-su-user = postgres
-postgres-password = aRWP1G52uu
-
-[API]
-mosip-api-internal-host = https://api-internal.dev1.mosip.net
-mosip_deployment_client_secret = ODlrq4kv1woCA6JR
