--- certalert/renew/expired.txt
@@ -1,1 +1,0 @@
-mpartner-default-print
