--- certalert/renew/requirements.txt
@@ -1,2 +1,0 @@
-psycopg2-binary==2.9.1
-requests==2.26.0
