+++ certmanager/bootstrap.properties
@@ -5,5 +5,5 @@ db-su-user = postgres
 postgres-password = cUZ5DLgLqz
 [API]
 mosip-api-internal-host=api-internal.soil-security.mosip.net
-mosip_deployment_client_secret=kxDGFoBJRsa9BprI
+mosip_pms_client_secret=kxDGFoBJRsa9BprI
 pre-expiry-days=400
