+++ certmanager/partner.properties
